move pushChanges into prismicRepository manager
